fix(name_resolution): rename resolvers to resolver
