+++ src/front/ast_retriever/name_resolution/scope_table.rs
@@ -1,5 +1,5 @@
-use crate::front::ast_retriever::name_resolution::resolvers::ResolveResult;
-use crate::front::ast_retriever::name_resolution::resolvers::ResolverError::Redefinition;
+use crate::front::ast_retriever::name_resolution::resolver::ResolveResult;
+use crate::front::ast_retriever::name_resolution::resolver::ResolverError::Redefinition;
 use crate::front::ast_types::{RawName, ResolvedName};
 use std::collections::HashMap;
 use std::rc::Rc;
